+++ src/assets/arrow.tsx
@@ -1,0 +1,15 @@
+const Arrow = () => {
+  return (
+    <svg width="25" height="25" viewBox="0 0 25 25" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <path
+        d="M9.5 15.5L3.5 9.5M3.5 9.5L9.5 3.5M3.5 9.5H15.5C17.0913 9.5 18.6174 10.1321 19.7426 11.2574C20.8679 12.3826 21.5 13.9087 21.5 15.5C21.5 17.0913 20.8679 18.6174 19.7426 19.7426C18.6174 20.8679 17.0913 21.5 15.5 21.5H12.5"
+        stroke="#7E8490"
+        stroke-width="1.5"
+        stroke-linecap="round"
+        stroke-linejoin="round"
+      />
+    </svg>
+  );
+};
+
+export default Arrow;
