+++ src/assets/four.tsx
@@ -8,10 +8,10 @@ const Four = () => {
       xmlns="http://www.w3.org/2000/svg"
     >
       <path
-        fillRule="evenodd"
-        clipRule="evenodd"
+        fill-rule="evenodd"
+        clip-rule="evenodd"
         d="M55.9493 136.135C55.7547 135.94 55.5644 135.742 55.3784 135.539C50.9131 130.668 48.7343 124.572 48.7343 118.35V112.3H25.4343C18.3187 112.3 12.4871 109.283 8.347 105.461C8.16094 105.29 7.97834 105.114 7.7993 104.935C3.27599 100.412 0.434326 94.2685 0.434326 87.3C0.434326 81.4401 2.43985 75.8589 6.23837 71.3473L53.7866 10.2351L53.7886 10.2326C56.1606 7.18371 59.5885 4.18516 64.2249 2.41491C67.1347 1.22753 70.3902 0.5 73.8843 0.5C80.5149 0.5 86.965 3.0433 91.8102 8.16057C96.6161 13.1266 98.7343 19.3974 98.7343 25.5V64.846C101.863 66.0507 104.766 67.912 107.269 70.415C112.045 75.1903 114.484 81.4217 114.484 87.9C114.484 94.6441 111.785 100.719 107.269 105.235C107.056 105.448 106.838 105.656 106.615 105.859C104.216 108.04 101.54 109.634 98.7343 110.691V118.35C98.7343 124.944 96.2426 131.112 91.654 135.846C86.8095 140.959 80.362 143.5 73.7343 143.5C67.0051 143.5 60.7194 140.905 55.9493 136.135ZM77.9343 122.7C79.1343 121.5 79.7343 120.05 79.7343 118.35V93.3H89.7843C91.3843 93.3 92.7343 92.8 93.8343 91.8C94.9343 90.7 95.4843 89.4 95.4843 87.9C95.4843 86.3 94.9343 84.95 93.8343 83.85C92.7343 82.75 91.3843 82.2 89.7843 82.2H79.7343V25.5C79.7343 23.8 79.1843 22.4 78.0843 21.3C77.6235 20.7973 77.11 20.3999 76.5439 20.1078C75.7985 19.7231 74.9617 19.5211 74.0337 19.5016C73.9842 19.5005 73.9344 19.5 73.8843 19.5C72.9843 19.5 72.0843 19.7 71.1843 20.1C70.2843 20.4 69.4843 21 68.7843 21.9L20.9343 83.4C20.0136 84.4128 19.5168 85.595 19.4438 86.9469C19.4375 87.0633 19.4343 87.181 19.4343 87.3C19.4343 87.85 19.5052 88.3764 19.647 88.8791C19.9177 89.8389 20.4468 90.7125 21.2343 91.5C22.5343 92.7 23.9343 93.3 25.4343 93.3H67.7343V118.35C67.7343 120.05 68.2843 121.5 69.3843 122.7C70.5843 123.9 72.0343 124.5 73.7343 124.5C75.4343 124.5 76.8343 123.9 77.9343 122.7ZM67.7343 82.2H35.6632L67.7343 40.3331V82.2Z"
-        fill="#AFDB00"
+        fill="#BEF264"
       />
     </svg>
   );
