+++ src/assets/index.ts
@@ -1,3 +1,4 @@
 export { default as Four } from './four';
 export { default as Four2 } from './four2';
 export { default as Zero } from './zero';
+export { default as Arrow } from './arrow';
